sync/requirements.txt
gspread==6.1.2
google-auth==2.29.0
requests==2.32.3
python-dotenv==1.2.1
